+++ lib/web/ntk_map_controller_web.dart
@@ -1,6 +1,11 @@
+import 'dart:js_interop_unsafe';
+
 import 'package:latlong2/latlong.dart';
+import 'package:ntk_map_view/models/map_button_class.dart';
 import 'package:ntk_map_view/models/map_filter_class.dart';
-import 'package:universal_html/js_util.dart' if (dart.library.io) '' as js_util;
+import 'package:ntk_map_view/models/map_marker_content_model.dart';
+import 'package:ntk_map_view/models/map_marker_icon_model.dart';
+import 'dart:js_interop' if (dart.library.io) '' as js_util;
 
 import '../interfaces/ntk_map_controller_interface.dart';
 import '../modules/create_unique_uid.dart';
@@ -16,22 +21,21 @@ class NtkMapController extends NtkMapControllerInterface {
       super.markers[point] = callback;
     }
 
-    js_util.callMethod<void>(
-        js_util.globalThis, '_addMarker', [point.latitude, point.longitude]);
+    js_util.globalContext.callMethodVarArgs('_addMarker'.toJS, [point.latitude.toJS, point.longitude.toJS]);
   }
 
   ///Move center camera on **[point]** (like panTo in leaflet)
   @override
   goToPoint(LatLng point) {
-    js_util.callMethod<void>(
-        js_util.globalThis, '_goToPoint', [point.latitude, point.longitude]);
+    js_util.globalContext.callMethodVarArgs('_goToPoint'.toJS, [point.latitude.toJS, point.longitude.toJS]);
+    // js_util.callMethod<void>(
+    //     js_util.globalThis, '_goToPoint', [point.latitude, point.longitude]);
   }
 
   ///Move camera to **[point]** and **[zoom]** (like flyTo in leaflet)
   @override
   goToPointThenZoom(LatLng point, double zoom) {
-    js_util.callMethod<void>(js_util.globalThis, '_goToPointThenZoom',
-        [point.latitude, point.longitude, zoom]);
+    js_util.globalContext.callMethodVarArgs('_goToPointThenZoom'.toJS, [point.latitude.toJS, point.longitude.toJS, zoom.toJS]);
   }
 
   ///Create marker on point with title, in this marker you may configure a button and its callback
@@ -39,37 +43,61 @@ class NtkMapController extends NtkMapControllerInterface {
   ///with **[title]**
   ///buttons with **[names]**
   ///and callbacks **[acts]**
+
   @override
   addCustomMarker(
-      LatLng point, String title, List<String> names, List<Function> acts) {
+      {required LatLng point,
+      required String title,
+        MapMarkerIconModel? icon,
+      required List<MapButtonClass> buttons,
+      List<String>? images,
+      List<MapMarkerContentModel>? content}) {
     List<String> buttIds = [];
-    for (var act in acts) {
-      String id = createUniqueUid(count: 6, isNumberEnabled: false);
 
+    for (int index = 0; index < buttons.length; index++) {
+      String id = createUniqueUid(count: 6, isNumberEnabled: false);
       buttIds.add(id);
+
+      buttons[index] = buttons[index]..id = id;
+
       //print(id);
-      super.markersAction[id] = act;
+      super.markersAction[id] = buttons[index].act;
     }
 
-    js_util.callMethod<void>(js_util.globalThis, '_addMarkerCustom',
-        [point.latitude, point.longitude, title, names, buttIds]);
+    js_util.globalContext.callMethodVarArgs('_addMarkerCustom'.toJS, [
+      point.latitude.toJS,
+      point.longitude.toJS,
+      title.toJS,
+      icon != null ? icon.toMapJs().toJS : "none".toJS,
+      [...buttons.map((e) => e.toMapJs().toJS)].toJS,
+      images != null ? [...images.map((e) => e.toJS)].toJS : null,
+      content != null ? [...content.map((e) => e.toMapJs().toJS)].toJS : null,
+    ]);
   }
 
   ///Create polyline on List of **[points]** (also clear all previous polyline)
   @override
   addPolyline(List<LatLng> points) {
-    List<List<double>> pointsForMap =
-        points.map((e) => [e.latitude, e.longitude]).toList();
+    List<List<double>> pointsForMap = points.map((e) => [e.latitude, e.longitude]).toList();
 
-    js_util.callMethod<void>(
-        js_util.globalThis, '_createPolyline', [pointsForMap]);
+    js_util.globalContext.callMethodVarArgs('_createPolyline'.toJS, [
+      [
+        ...pointsForMap.map((e) => [...e.map((last) => last.toJS)].toJS)
+      ].toJS
+    ]);
+
+    // js_util.callMethod<void>(
+    //     js_util.globalThis, '_createPolyline', [pointsForMap]);
   }
 
   ///Update **[filter]** map
   @override
   applyNewFilter(MapFilter filter) {
-    js_util.callMethod<void>(
-        js_util.globalThis, '_updateFilter', [filter.toParameterString()]);
+    js_util.globalContext.callMethodVarArgs('_updateFilter'.toJS, [
+      [...filter.toParameterString().map((e) => e.toJS)].toJS
+    ]);
+    // js_util.callMethod<void>(
+    //     js_util.globalThis, '_updateFilter', [filter.toParameterString()]);
   }
 
   ///Update current position on map
@@ -77,7 +105,16 @@ class NtkMapController extends NtkMapControllerInterface {
   ///Radius of circle is **[accuracy]**
   @override
   updateCurrentPosition(LatLng point, double accuracy) {
-    js_util.callMethod<void>(js_util.globalThis, '_updateCurrentPosition',
-        [point.latitude, point.longitude, accuracy]);
+    js_util.globalContext.callMethodVarArgs('_updateCurrentPosition'.toJS, [point.latitude.toJS, point.longitude.toJS, accuracy.toJS]);
+
+    // js_util.callMethod<void>(js_util.globalThis, '_updateCurrentPosition',
+    //     [point.latitude, point.longitude, accuracy]);
+  }
+
+  @override
+  clearAllMarkers() {
+    js_util.globalContext.callMethod(
+      '_clearAllMarkers'.toJS,
+    );
   }
 }